ENH: Add private/public/desc studies view
Private studies can now be viewed from the Study tab on the top toolbar
and from there the user can navigate to see a description of the study.
Note that currently the study description is not finished therefore the
content there presented must be changed.

One subtle change that took place during in this changeset is moving the
name "my studies" to "private studies" to make this semantically
consistent.