TST: Fix tests for template serialization
Two big changes happened, the column that contains the sample names
changed from being sample_name to being, the QIIME standar, #SampleID,
the test strings had to be updated to reflect this.

Additionally the field ebi_study_accession was removed from the prep
templates so I updated this in the test string.