BUG: Fix incorrect investigation type for studies
All the studies that we are currently processing are amplicon sequencing
and thus not defined in the ontology given by EBI, see link below. Since
we knew this, the objects and underlying code submit as "Other" and with
a value for "new_study_type". However the GUI only allowed you to select
from one of the choices that EBI provides.

The problem is explained in more detail in #522 and once that PR is
merged it should be fixed, this change makes it so that the submissions
include correct information.

https://www.ebi.ac.uk/ega/sites/ebi.ac.uk.ega/files/documents/Study.xml