MAINT:Change ontology term identifier to bigserial
The term identifier in the ontology terms table was set as a big int,
thus adding new user-defined terms was cumbersome, with this change in
place those the term is now a bigserial.

I have also changed the ontology file in the beast, so people WILL HAVE
to reload their databases.