BUG: Fix smaple_names being casted to int/float
read_csv was casting the `'sample_name'` column to a float if the values
were of the form `0000012.122204`, this has been fixed and we have also
added a new keyword argument to skip lines that are "empty" or that lack
a sample_name i. e. if a line starts with a `\t` character it will be
ignored. Empty columns are also now removed.