ENH: Uploads are no longer concatenated
The uploads handler appends every chunk of transmitted data into an
"in-construction" file. This avoids the overhead caused by first
writing all the chunks and then iterating over every chunk to construct
the final file.

An important thing to note with this patch is that uploads will not be
resumable as chunks are appended into the final file as they come and
thus there's no way to figure out which chunks were successfully
transmitted and which weren't.

Fixes #599