BUG:Remove rows without sample name & improve docs
Rows with empty sample names are now removed from the dataframe, note
that these rows could include relevant data in other columns but since
we have no sample name, they are useless.

The documentation for read_csv was lacking a bit, so I expanded on the
reason why the arguments are being passed the way they are.

Fixes #746
Fixes #757