ENH/API: methods to update sample metadata
These changes include:

* updating of existing sample metadata
* adding new columns
* removing columns

Right now, this is limited to the SampleTemplate, and support is not in
for the PrepTemplate. This would be easy to change, but put it in place
this way based on the email traffic I saw regarding the issue.

The API for get/set/del at the template level is not consistent yet. Pushing
in now to exercise full tests, and will issue follow up commits to resolve the
inconsistencies.