Merged BIOM post-processing
analysis.build_files will now run a series of scripts to the merged
biom, according to the parameters stored in qiita.software_command.
post_processing_cmd. Currently, the merged biom is passed as a
parameter, and assumes that each process in turn will overwrite
the merged biom file as the method to return the result back to the
user.
A trivially sized script named 'worker.py' allows the unit tests to
run a sample script and confirm the environment that it is running
in.