Removing dead file
This file contains a comment at the bottom stating that it cannot be
called directly, and should only be used as a reference to help recover
jobs. Since this file is possibly stale, and will definitely be
out-of-date after this change is complete, there is no reason to keep
it.