Sanitize 'Add Person' name
In Create/Edit Study page, there is an option to add PIs and staff using
pull-down menus. New PIs and staff are added to the system using a
sub-form. This sub-form has a field for the person's name. This field
cannot have non-printable ASCII characters, as the PI name becomes part
of system fields downstream in Qiita. This will not correct names for
existing persons entered, only sanitize names for future persons.