Study_Person's name can now only contain ASCII
A person's name in qiita.study_person can now only contain printable
ASCII characters, as this information is used in the construction of
other generated names. These names are used in places that are sensitive
to non-ASCII characters.

Updated commit includes a description message for the user, and a patch
for the DB to strip characters from existing entries.