Fixes Issue 2768 (#2771)
* Sanitize 'Add Person' name

In Create/Edit Study page, there is an option to add PIs and staff using
pull-down menus. New PIs and staff are added to the system using a
sub-form. This sub-form has a field for the person's name. This field
cannot have non-printable ASCII characters, as the PI name becomes part
of system fields downstream in Qiita. This will not correct names for
existing persons entered, only sanitize names for future persons.

* Study_Person's name can now only contain ASCII

A person's name in qiita.study_person can now only contain printable
ASCII characters, as this information is used in the construction of
other generated names. These names are used in places that are sensitive
to non-ASCII characters.

Updated commit includes a description message for the user, and a patch
for the DB to strip characters from existing entries.

* Update 68.sql