Added support for validator process chains
When using Torque as job submission system, Qiita will now organize an
artifact's m validator jobs into n 'processing chains', where a job
will only be run if its predecessor in the list completed successfully.

Currently, the number of processing jobs an artifact can have running
will be two. This is currently configurable in the source.

As a consequence, Launchers have been modified to be able to execute
either in-process or forked() using Process(). When running jobs
locally, job creation will behave as previously, with a forked Process()
waiting on the job to finish, and subsequently updating the database.

When submitting jobs to Torque, the launcher will execute the command
inside submit(), expecting it to return quickly and with a job id. This
id can them be used when creating validator jobs to organize them into n
'queues'.

A new class, tentatively named Watcher, is expected to be run as a
separate Thread() or Process() within the main Qiita script. It will
update Qiita when a job has completed with the correct exit status and
error message, if any.