Fix qsub hanging error (#2796)
* Fix qsub hanging error

It appears that qsub can hang when not given a config file, or perhaps a
parameter is in error; it assumes that more information will be provided
on stdin, hence it will wait forever. If this happens within Tornado, it
will cause Qiita workers to become unresponsive.

Python3 allows one to call .communicate() with a timeout. However, this
feature is not available in Python2. As a temporary measure, Qopen has
been created to wrap Popen() calls in a different Thread, and the Qopen
object can be subsequently joined with a timeout.

* Fix pip to version 18.1

* Fix pip=18.1 to pip==18.1