/api/v1/study/1/samples accepts new categories
calling PATCH on /api/v1/study/1/samples now silently accepts metadata
with new categories in them when updating a sample or adding a new one.
Existing samples will be assigned a null value while samples in the
metadata will be assigned the value or empty string if one was not
given.